scripts/requirements.txt
distro==1.9.0
httplib2==0.22.0
launchpadlib==1.11.0
lazr-restfulclient==0.14.6
lazr-uri==1.0.6
semver==3.0.2
six==1.16.0
wadllib==1.3.6
